Added some controllers and cleaned up simulink project
A basic set of controllers, ranging from LQR to MRAC to SMC will be implemented
for the considered processes, both in discrete and continuous time. So far, I've merely collected the code from old projects and cleaned it up a bit. The
standard DC servo is supported with various controllers. Next up is the Furuta pendulum.